requirements.txt
streamlit
pandas
numpy
plotly
pdfplumber
graphviz
biopython
transformers
torchss